change display on quiz result components, and split QuestionScreen into components
