+++ package.json
@@ -26,9 +26,12 @@
     "expo-location": "^18.0.10",
     "expo-status-bar": "~2.0.1",
     "geolib": "^3.3.4",
+    "lottie-react-native": "^7.2.2",
     "lucide-react-native": "^0.488.0",
     "react": "18.3.1",
     "react-native": "0.76.9",
+    "react-native-gifted-charts": "^1.4.60",
+    "react-native-linear-gradient": "^2.8.3",
     "react-native-maps": "1.18.0",
     "react-native-safe-area-context": "^5.3.0",
     "react-native-screens": "^4.10.0",
